:recycle: Closing JYXB-132 UseCase 및 Service 구현체 분리
