+++ src/main/java/com/promptoven/authservice/adaptor/web/controller/MemberRestController.java
@@ -1,12 +1,15 @@
 package com.promptoven.authservice.adaptor.web.controller;
 
+import com.promptoven.authservice.adaptor.web.controller.vo.in.UpdateNicknameRequestVO;
+import com.promptoven.authservice.application.port.in.usecase.MemberManagementUseCase;
+
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.promptoven.authservice.adaptor.web.controller.vo.in.UpdateNicknameRequestVO;
-import com.promptoven.authservice.application.port.in.usecase.MemberUseCases;
+
+
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -17,7 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 @RequestMapping("/v1/auth/member")
 public class MemberRestController {
 
-	private final MemberUseCases memberUseCases;
+	private final MemberManagementUseCase memberUseCases;
 
 	@PutMapping("/nickname")
 	public void updateNickname(@RequestBody UpdateNicknameRequestVO updateNicknameRequestVO) {
