+++ src/main/java/com/promptoven/authservice/application/port/in/usecase/MediaAuthUseCase.java
@@ -1,0 +1,11 @@
+package com.promptoven.authservice.application.port.in.usecase;
+
+public interface MediaAuthUseCase {
+
+    boolean checkMedia(String email, String code);
+
+    void requestEmail(String email);
+
+    void requestPhone(String phone);
+
+}
