+++ src/main/java/com/promptoven/authservice/application/port/in/usecase/MemberManagementUseCase.java
@@ -1,0 +1,16 @@
+package com.promptoven.authservice.application.port.in.usecase;
+
+public interface MemberManagementUseCase {
+
+    void promoteToSeller(String memberUUID);
+
+    void setMemberRole(String memberUUID, String roleName);
+
+    void banMember(String memberUUID);
+
+    void unbanMember(String memberUUID);
+
+    void updateNickname(String memberUUID, String nickname);
+
+    void clearPassword(String memberUUID);
+}
